refactor: move ResourceInfo to runtime module
